WIP: Generalization of the 'include' config section
This is a generalization of the spack-env's 'include:' feature into a
new 'include' config section.

So config scopes can now have their own 'include.yaml'. And spack-env's
'spack:include:' works same as before, but is now seen as a config
section (like 'packages:') instead of being spack-env-specific thing.

The current implementation is very similar to the one that were written
for spack-envs: when registering a new config scope in the
`spack.config.config.scopes` list, the 'include' config section is read
and scopes are pushed recursively (at lower precedence than the
includer). The `remove_scope` implementation is trickier, but works as a
POC for now.