Environments: Avoid inconsistent state on failed write (#18538)
Fixes #18441

When writing an environment, there are cases where the lock file for
the environment may be removed. In this case there was a period
between removing the lock file and writing the new manifest file
where an exception could leave the manifest in its old state (in
which case the lock and manifest would be out of sync).

This adds a context manager which is used to restore the prior lock
file state in cases where the manifest file cannot be written.

(cherry picked from commit 8ad2cc2acfc2374c9b6bc01cd2f1c8eb97f7c7f0)